Update main.c
changing all data types in addition to float

+++ main.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 float subtract(float num1,float num2);
-int add(int num1,int num2);
+float add(float num1,float num2);
 float multiplication(float num1, float num2);
 float division(float num1, float num2);
